fix: change configs and eslint rule
