+++ package.json
@@ -26,6 +26,7 @@
     "@types/react-dom": "^18.3.0",
     "@types/react-redux": "^7.1.33",
     "@types/react-router-dom": "^5.3.3",
+    "@types/react-test-renderer": "^18.3.0",
     "@vitejs/plugin-react": "^4.2.1",
     "babel-jest": "^29.7.0",
     "eslint": "^8.57.0",
@@ -58,6 +59,7 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^2.2.3",
+    "@types/react-router": "^5.1.20",
     "axios": "^1.6.8",
     "classnames": "^2.5.1",
     "react": "^18.3.1",
